bigquery/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
flask
